+++ .reuse/dep5
@@ -14,3 +14,7 @@ License: CC-BY-4.0
 Files: *.png
 Copyright: 2024 Phoenix R&D GmbH <hello@phnx.im>
 License: CC-BY-4.0
+
+Files: prototype/rust_builder/cargokit/build_tool/lib/*
+Copyright: 2024 Phoenix R&D GmbH <hello@phnx.im>
+License: AGPL-3.0-or-later
