+++ prototype/integration_test/simple_test.dart
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2024 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:prototype/main.dart';
 import 'package:prototype/core/frb_generated.dart';
