--- prototype/rust
@@ -1,1 +1,0 @@
-../applogic
